Move runCom1DFA scripts
